Remove unused code sections
Removed functions alt_ping() and ctrl_alt_ping(), as well as the
loading of bumblebee.mp3 that was commented out.

alt_ping() and ctrl_alt_ping() aimed to simulate a left-click with
the alt (or alt and ctrl) keys held down, i.e.
    pyautogui.keyDown('alt')
    pyautogui.click()
    pyautogui.keyUp('alt')
However, this does not seem to equate to a ping click in Dota 2.
I'm not sure why exactly, though the following comment from
pyautogui's __init__.py may be of relevance:

> NOTE: For some reason, this does not seem to cause key repeats
> like would happen if a keyboard key was held down on a text
> field.

bumblebee.mp3 was removed to simplify the script (keeping it
minimal). I may consider a try/except statement in loading
bumblebee.mp3 in the future.