Tweak key/mouse input functions into class with decorators
Organised the left click, right click, alt ping, and ctrl alt ping
functions into a class, with the use of @alt and @ctrl decorators.

I was unable to organise the pygame.mixer.music functions into a
class of their own: for some reason, this version of the code with
the pygame functions within a class does not work at all. Outside,
however, it functions as intended.